requirements.txt
# FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9  # For enhanced form data parsing (NLP feature)

# Pydantic (Data Validation)
pydantic==2.9.2
pydantic-settings==2.6.1

# Database
psycopg2-binary==2.9.10
pgvector==0.3.6

# ML/AI - Sentence Transformers
sentence-transformers==3.3.1
torch==2.5.1  # ✅ Latest stable for Apple Silicon (macOS ARM64)
torchvision==0.20.1  # ✅ Compatible with torch 2.5.1
transformers==4.46.3
numpy==1.26.4

# Utilities
python-dotenv==1.0.1
redis==5.2.0
httpx==0.27.2
nltk==3.9.1  # Natural Language Toolkit for intelligent synonym expansion

# Development (Optional)
pytest==8.3.3
pytest-asyncio==0.24.0